chore: add onContextMenu event
